requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.1
    # via requests
django==5.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
idna==3.10
    # via requests
requests==2.32.3
    # via -r requirements.in
sqlparse==0.5.3
    # via django
urllib3==2.4.0
    # via requests
